split logic into seperate files
